+++ src/core/store/wallet/getters.js
@@ -21,8 +21,13 @@ const tokensList = function (state) {
     : [];
 };
 
+const initialLoad = function (state) {
+  return state.initialLoadTokens || state.initialLoadBalance;
+};
+
 export default {
   balanceInETH,
   totalOwnedDomains,
-  tokensList
+  tokensList,
+  initialLoad
 };